+++ packages/badge/rollup.config.ts
@@ -1,3 +1,3 @@
-import { defineConfig } from '@badge-collection/rollup'
+import { defineConfig } from '@readme-widget-hub/rollup'
 
 export default defineConfig()