+++ packages/meta/CHANGELOG.md
@@ -1,4 +1,15 @@
-# @badge-collection/meta
+# @readme-widget-hub/meta
+
+## 2.0.0
+
+### Major Changes
+
+- 项目改名为 Readme Widget Hub
+
+### Patch Changes
+
+- Updated dependencies
+  - @readme-widget-hub/utils@2.0.0
 
 ## 1.3.0
 
@@ -27,4 +38,4 @@
 ### Patch Changes
 
 - Updated dependencies
-  - @badge-collection/utils@1.0.0
+  - @readme-widget-hub/utils@1.0.0