+++ packages/utils/CHANGELOG.md
@@ -1,4 +1,10 @@
-# @badge-collection/utils
+# @readme-widget-hub/utils
+
+## 2.0.0
+
+### Major Changes
+
+- 项目改名为 Readme Widget Hub
 
 ## 1.0.0